refactor: member 중복 생성 로직 제거
