+++ src/main/java/roomescape/member/service/MemberService.java
@@ -1,15 +1,14 @@
 package roomescape.member.service;
 
-import java.util.List;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import roomescape.auth.controller.dto.SignUpRequest;
 import roomescape.exception.ErrorType;
 import roomescape.exception.NotFoundException;
-import roomescape.member.controller.dto.MemberResponse;
+import roomescape.auth.controller.dto.MemberResponse;
 import roomescape.member.domain.Member;
-import roomescape.member.domain.Role;
 import roomescape.member.domain.repository.MemberRepository;
+
+import java.util.List;
 
 @Service
 @Transactional(readOnly = true)
@@ -24,13 +23,6 @@ public class MemberService {
         return memberRepository.findAll().stream()
                 .map(MemberResponse::from)
                 .toList();
-    }
-
-    @Transactional
-    public MemberResponse create(SignUpRequest signUpRequest) {
-        Member member = memberRepository.save(
-                new Member(signUpRequest.name(), signUpRequest.email(), signUpRequest.password(), Role.USER));
-        return new MemberResponse(member.getId(), member.getName());
     }
 
     public Member findById(Long memberId) {
