+++ src/main/java/roomescape/reservation/controller/dto/ReservationResponse.java
@@ -2,7 +2,7 @@ package roomescape.reservation.controller.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import java.time.LocalDate;
-import roomescape.member.controller.dto.MemberResponse;
+import roomescape.auth.controller.dto.MemberResponse;
 import roomescape.member.domain.Member;
 import roomescape.reservation.domain.MemberReservation;
 import roomescape.reservation.domain.Reservation;
