+++ src/test/java/roomescape/member/service/MemberServiceTest.java
@@ -6,8 +6,6 @@ import static roomescape.fixture.MemberFixture.getMemberClover;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import roomescape.auth.controller.dto.SignUpRequest;
-import roomescape.member.controller.dto.MemberResponse;
 import roomescape.member.domain.Member;
 import roomescape.member.domain.repository.MemberRepository;
 import roomescape.util.ServiceTest;
@@ -18,21 +16,6 @@ class MemberServiceTest extends ServiceTest {
     MemberRepository memberRepository;
     @Autowired
     MemberService memberService;
-
-    @DisplayName("사용자 생성에 성공한다.")
-    @Test
-    void create() {
-        //given
-        String password = "1234";
-        Member memberClover = getMemberClover();
-        SignUpRequest signUpRequest = new SignUpRequest(memberClover.getName(), memberClover.getEmail(), password);
-
-        //when
-        MemberResponse memberResponse = memberService.create(signUpRequest);
-
-        //then
-        assertThat(memberResponse.name()).isEqualTo(memberClover.getName());
-    }
 
     @DisplayName("식별자로 사용자 조회에 성공한다.")
     @Test
